+++ app/src/main/java/com/datnt/textart/fragment/create/RecentFragment.java
@@ -1,18 +1,13 @@
 package com.datnt.textart.fragment.create;
 
 import android.content.Intent;
-import android.graphics.Bitmap;
-import android.net.Uri;
 import android.os.Bundle;
 
 import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.GridLayoutManager;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
-import androidx.viewpager2.widget.ViewPager2;
 
-import android.provider.MediaStore;
-import android.util.Base64;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -29,8 +24,6 @@ import com.datnt.textart.model.BucketPicModel;
 import com.datnt.textart.model.PicModel;
 import com.datnt.textart.sharepref.DataLocalManager;
 
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
 import java.util.ArrayList;
 
 public class RecentFragment extends Fragment {
@@ -64,7 +57,7 @@ public class RecentFragment extends Fragment {
         rlExpand = view.findViewById(R.id.rlExpand);
         vBg = view.findViewById(R.id.viewBg);
 
-        rlExpand.getLayoutParams().height = (int) getResources().getDisplayMetrics().heightPixels * 60 / 100;
+        rlExpand.getLayoutParams().height = getResources().getDisplayMetrics().heightPixels * 60 / 100;
 
         setUpLayout();
         evenClick();
